# Conflicts:
#	app/src/main/java/dk/eboks/app/network/util/ItemTypeAdapterFactory.kt
#
app/src/main/java/dk/eboks/app/presentation/ui/components/channels/openi
ng/ChannelOpeningComponentFragment.kt

+++ app/src/main/java/dk/eboks/app/domain/interactors/message/GetReplyFormInteractor.kt
@@ -11,7 +11,7 @@ interface GetReplyFormInteractor : Interactor {
     var output : Output?
     var input : Input?
 
-    data class Input(var messageId: String, var folderId: Long)
+    data class Input(var messageId: String, var folderId: Int)
 
     interface Output {
         fun onGetReplyForm(form : ReplyForm)
